refactor(xod-client): implement undo/redo usind reducer enhancer instead of middleware
