+++ packages/xod-client/src/core/keepIntegrityAfterNavigatingHistory.js
@@ -4,11 +4,6 @@ import { selectionLens } from '../editor/selectors';
 import { SELECTION_ENTITY_TYPE } from '../editor/constants';
 import { getCurrentPatchLinks, getCurrentPatchNodes } from '../project/selectors';
 
-import {
-  PATCH_HISTORY_UNDO,
-  PATCH_HISTORY_REDO,
-} from '../project/actionTypes';
-
 const removeInvalidSelections = (state) => {
   const nodes = getCurrentPatchNodes(state);
   const links = getCurrentPatchLinks(state);
@@ -28,12 +23,4 @@ const removeInvalidSelections = (state) => {
   )(state);
 };
 
-export default (state, action) => {
-  switch (action.type) {
-    case PATCH_HISTORY_UNDO:
-    case PATCH_HISTORY_REDO:
-      return removeInvalidSelections(state);
-    default:
-      return state;
-  }
-};
+export default removeInvalidSelections;
