+++ packages/xod-client/src/core/middlewares.js
@@ -2,9 +2,8 @@ import { compose, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
 import { devToolMiddleware } from '../utils/devtools';
-import projectHistoryMiddleware from '../project/projectHistoryMiddleware';
 
 export default compose(
-  applyMiddleware(thunk, projectHistoryMiddleware),
+  applyMiddleware(thunk),
   devToolMiddleware
 );
