+++ packages/xod-client/src/core/reducer.js
@@ -2,7 +2,7 @@ import { merge } from 'ramda';
 import { combineReducers } from 'redux';
 
 import projectReducer from '../project/reducer';
-import historyReducer from './historyReducer';
+import undoableProject from './undoableProject';
 import projectBrowserReducer from '../projectBrowser/reducer';
 import editorReducer from '../editor/reducer';
 import errorsReducer from '../messages/reducer';
@@ -10,11 +10,6 @@ import processesReducer from '../processes/reducer';
 import popupsReducer from '../popups/reducer';
 
 import keepIntegrityAfterNavigatingHistory from './keepIntegrityAfterNavigatingHistory';
-
-const composeReducers =
-  (...reducers) =>
-    (state, action) =>
-      reducers.reduceRight((st, r) => r(st, action), state);
 
 const combineRootReducers = (extraReducers) => {
   const reducers = merge(
@@ -30,10 +25,9 @@ const combineRootReducers = (extraReducers) => {
     extraReducers
   );
 
-  return composeReducers(
+  return undoableProject(
     combineReducers(reducers),
-    keepIntegrityAfterNavigatingHistory,
-    historyReducer
+    keepIntegrityAfterNavigatingHistory
   );
 };
