+++ packages/xod-client/src/project/actionTypes.js
@@ -21,6 +21,3 @@ export const PATCH_DELETE = 'PATCH_DELETE';
 
 export const PATCH_HISTORY_UNDO = 'PATCH_HISTORY_UNDO';
 export const PATCH_HISTORY_REDO = 'PATCH_HISTORY_REDO';
-export const PATCH_HISTORY_SAVE = 'PATCH_HISTORY_SAVE';
-export const PATCH_HISTORY_CLEAR_FOR_PATCH = 'PATCH_HISTORY_CLEAR_FOR_PATCH';
-export const PATCH_HISTORY_CLEAR_ALL = 'PATCH_HISTORY_CLEAR_ALL';
